# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "User Service.",
        "title": "User Service",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/v1/payroll/run": {
            "post": {
                "description": "Admin runs payroll for a given attendance period",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payroll"
                ],
                "summary": "Run payroll",
                "parameters": [
                    {
                        "description": "RunPayroll payload",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.PayrollRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.PayrollResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.PayrollResponse": {
            "type": "object",
            "properties": {
                "attendance_period_id": {
                    "type": "string"
                },
                "payroll_id": {
                    "type": "string"
                },
                "run_at": {
                    "type": "string"
                }
            }
        },
...
        "request.PayrollRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "period_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}